# idm/generated/openapi/src/main/java/com/onyxdb/idm/generated/openapi/apis/ProjectsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.5.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.onyxdb.idm.generated.openapi.apis;

import com.onyxdb.idm.generated.openapi.models.BadRequestResponse;
import com.onyxdb.idm.generated.openapi.models.NotFoundResponse;
import com.onyxdb.idm.generated.openapi.models.ProjectDTO;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.5.0")
@Validated
@Tag(name = "Projects", description = "the Projects API")
public interface ProjectsApi {

    /**
     * POST /api/v1/projects : Create a new project
     *
     * @param projectDTO  (required)
     * @return Created (status code 201)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "createProject",
        summary = "Create a new project",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjectDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/v1/projects",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProjectDTO> createProject(
        @Parameter(name = "ProjectDTO", description = "", required = true) @Valid @RequestBody ProjectDTO projectDTO
    );


    /**
     * DELETE /api/v1/projects/{projectId} : Delete a project by ID
     *
     * @param projectId  (required)
     * @return No Content (status code 204)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "deleteProject",
        summary = "Delete a project by ID",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "204", description = "No Content"),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/v1/projects/{projectId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> deleteProject(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId
    );


    /**
     * GET /api/v1/projects : Get all projects
     *
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getAllProjects",
        summary = "Get all projects",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = ProjectDTO.class)))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/projects",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<ProjectDTO>> getAllProjects(
        
    );


    /**
     * GET /api/v1/projects/{projectId} : Get a project by ID
     *
     * @param projectId  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "getProjectById",
        summary = "Get a project by ID",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjectDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/v1/projects/{projectId}",
        produces = { "application/json" }
    )
    
    ResponseEntity<ProjectDTO> getProjectById(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId
    );


    /**
     * PUT /api/v1/projects/{projectId} : Update a project by ID
     *
     * @param projectId  (required)
     * @param projectDTO  (required)
     * @return OK (status code 200)
     *         or Not Found (status code 404)
     *         or Bad Request (status code 400)
     */
    @Operation(
        operationId = "updateProject",
        summary = "Update a project by ID",
        tags = { "Projects" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProjectDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Not Found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = NotFoundResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = BadRequestResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/v1/projects/{projectId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<ProjectDTO> updateProject(
        @Parameter(name = "projectId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("projectId") UUID projectId,
        @Parameter(name = "ProjectDTO", description = "", required = true) @Valid @RequestBody ProjectDTO projectDTO
    );

}